Add PUT, POST, PATCH, and DELETE
This commit adds the other request types that were needed for the library to
move forward. It shifts a few things around to get it to work but it works now.
It had to make the return type change to `(StatusCode, Option<Json>)` to work
as I found I was causing unwrapping on empty bodies to happen with things that
aren't returned. This also updates the accompanying docs/examples to work.

Either way maintainers now can implement any endtype without fear. It's now just
implementing all the endpoints!