Add ability to use custom endpoint for requests
It might be a while to get the library to a completely usable state. In the
meantime we should still allow users to be able to use endpoints not covered in
the code base. On top of that some users might need access to hidden endpoints
and those are not publically available. This is the solution for these problems.
While not foolproof since users might mistype the string being used as input it
still allows flexibility.

The main draw is that users can still use this API as is and as patches come out
adding new endpoints they can move the code over to that.