Switch Core to Rc<RefCell<Core>> in Github Struct
Before we had no good way to expose the event loop to users if they needed it
because doing so would violate the ownership rules. On top of that the lifetime
annotations and problems with the old implementation were a hastle to deal with.
If we limit the need for mutability with a user they have more flexibility due
to the imutability, especially with this API.