Have Client be internal to the Github struct
While looking through the Hyper issues I found out that creating a client each
time was a really bad thing to do. This commit changes it so that it's created
once and then referenced when crafting a request for the event loop upon
execution rather than being made each time.