Change to Hyper v0.11
This knocks off the dependency of the git branch of Hyper which is
great. However, a release cannot be made till a new release of hyper-tls
or another tls adaptor for Hyper is released. When that's done 0.5 can
finally be released.

There is a breaking change now that I had intended to make before. The `new`
function for `Github` type returns a `Result` now. You'll need to handle
that error now when creating it.