Adopt Rustfmt Style (?) (#166)
* fmt: Add a blank .rustfmt.toml file

(I will add rules to this file later---for now, simply add the file so
it is checked in.)

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>

* fmt: Re-organize 'extern crate', 'use', 'pub mod' statements

In accordance with rustfmt's default, this commit reorders 'extern
crate', 'use', and 'pub mod' statements to be in a consistent
order. (namely, alphabetical)

This should not have any effect on the compiled results, but makes the
codebase more consistent.  This is the default behavior that rustfmt
uses.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>

* fmt: Remove random newlines

This is not consistently done throughout the codebase, and rustfmt
removes these by default.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>

* fmt: Ensure newlines present at end of file

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>

* fmt: Ensure space between macro name and body

For consistency, and also because this is rustfmt's default behavior.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>

* fmt: Apply default rustfmt style

This is done to establish a baseline for what a "consistent" style
looks like.  (Whether or not this style is preferred is up for debate,
but this is the default style that rustfmt enforces.)  Stylistic
changes from here can be made easily by changing settings in the
.rustfmt.toml file.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>

* Delete .rustfmt.toml

We don't need this unless we customize the behavior, which we are not
going to do.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>